+++ core/src/com/mygdx/game/Vector.java
@@ -26,11 +26,11 @@ public class Vector {
         return this.z;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
